# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: ProxySieve local control-plane API
  version: 0.1.0
  description: >-
    The authenticated local Admin API exposed by ProxySieve. This contract
    describes the implemented surface; planned resources are intentionally omitted.
servers:
  - url: http://127.0.0.1:9090
    description: Default local-only listener
...
paths:
...
  /api/v1/chains:
    get:
      tags: [Chains]
      parameters: [{$ref: '#/components/parameters/Limit'}, {name: after, in: query, schema: {type: string}}]
      responses:
        '200': {description: Revisioned ordered proxy-chain inventory, content: {application/json: {schema: {$ref: '#/components/schemas/ChainPage'}}}}
    post:
      tags: [Chains]
      security: [sessionCookie: [], csrfToken: []]
      requestBody:
        required: true
        content: {application/json: {schema: {type: object, required: [chain], properties: {chain: {$ref: '#/components/schemas/ChainInput'}}}}}
      responses:
        '201': {description: Chain inventory record created, content: {application/json: {schema: {$ref: '#/components/schemas/ChainResponse'}}}}
        '400': {$ref: '#/components/responses/BadRequest'}
        '409': {description: Chain ID already exists}
...
  /api/v1/chains/{chainId}/test:
    parameters: [{name: chainId, in: path, required: true, schema: {type: string}}]
    post:
      tags: [Chains]
      security: [sessionCookie: [], csrfToken: []]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [target_host, target_port]
              properties:
                target_host: {type: string, minLength: 1, maxLength: 253}
                target_port: {type: integer, minimum: 1, maximum: 65535}
      responses:
        '200': {description: Sanitized connectivity result for the exact active chain revision, content: {application/json: {schema: {$ref: '#/components/schemas/ChainTestResponse'}}}}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {description: Chain not found}
        '409': {description: Saved chain does not exactly match the active runtime revision}
        '503': {description: Chain testing unavailable}
...
components:
  securitySchemes:
    sessionCookie: {type: apiKey, in: cookie, name: proxysieve_session}
    csrfToken: {type: apiKey, in: header, name: X-CSRF-Token}
...
  parameters:
    Limit: {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 1000, default: 100}}
...
  responses:
    BadRequest: {description: Request was not accepted, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: {type: object, required: [code, message], properties: {code: {type: string}, message: {type: string}}}
...
    ChainHop:
      type: object
      required: [pool_id, timeout_ns]
      properties:
        pool_id: {type: string}
        timeout_ns: {type: integer, minimum: 0, maximum: 120000000000}
    ChainInput:
      type: object
      description: Ordered chain metadata accepted on create; omit id to let the server generate it.
      required: [name, hops, enabled]
      properties:
        id: {type: string}
        name: {type: string, minLength: 1, maxLength: 256}
        hops: {type: array, minItems: 2, maxItems: 8, items: {$ref: '#/components/schemas/ChainHop'}}
        enabled: {type: boolean}
    Chain:
      allOf:
        - {$ref: '#/components/schemas/ChainInput'}
        - {type: object, required: [id]}
    ChainRecord:
      type: object
      required: [chain, revision]
      properties: {chain: {$ref: '#/components/schemas/Chain'}, revision: {type: integer, minimum: 1}}
    ChainResponse:
      type: object
      required: [chain, runtime_active, activation, runtime_revision]
      properties: {chain: {$ref: '#/components/schemas/ChainRecord'}, runtime_active: {type: boolean}, activation: {type: string, enum: [active, staged]}, runtime_revision: {type: integer, minimum: 0}, health: {$ref: '#/components/schemas/ChainHealth'}}
    ChainPage:
      type: object
      required: [items, next_after]
      properties: {items: {type: array, items: {$ref: '#/components/schemas/ChainResponse'}}, next_after: {type: string}}
    ChainHealth:
      type: object
      required: [status]
      properties:
        status: {type: string, enum: [untested, healthy, unhealthy]}
        tested_at: {type: string, format: date-time}
        latency_ns: {type: integer, minimum: 0}
        failure_reason: {type: string, enum: [route_unavailable, connect_failed]}
        failed_hop: {type: integer, minimum: 1, maximum: 8}
        pool_id: {type: string}
        proxy_id: {type: string}
    ChainTestResponse:
      type: object
      required: [result]
      properties:
        result: {$ref: '#/components/schemas/ChainHealth'}
...